Edit calendar now rerender when changes
