--- front-end/src/components/UpdateEvent/update.css
@@ -1,53 +1,0 @@
-.update-container{
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  width: 100%;
-}
-
-.update-button{
-  margin: 0px 60px;
-  max-width: 100%;
-  height: 35px;
-  max-width: 450px;
-  background-color: rgb(230, 230, 230);
-  border: 1px solid rgb(172, 172, 172);
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.update-container .event-container{
-  display: flex;
-  background-color: rgb(230 230 230);
-  max-width: 450px;
-  margin: 50px 60px 20px 60px;
-  height: 280px;
-  flex-direction: column;
-  padding: 36px;
-  border: 1px solid rgb(172, 172, 172);
-  border-radius: 5px;
-}
-
-.update-container .event-container .line{
-  max-width: 100%;
-  height: 2px;
-  margin: 5px 20px;
-  background-color: black;
-}
-
-.update-container .event-container .date-inputs{
-  display: flex;
-  flex-direction: column;
-}
-
-.delete-button{
-  margin: 0px 60px;
-  max-width: 100%;
-  height: 35px;
-  max-width: 450px;
-  background-color: rgb(255, 127, 127);
-  border: 1px solid rgb(172, 172, 172);
-  border-radius: 5px;
-  cursor: pointer;
-  margin-top: 15px;
-}
